requirements.txt
azure-functions>=1.17
azure-identity>=1.15
azure-cosmos>=4.5
msgraph-sdk>=1.2
litellm>=1.30
semantic-kernel>=1.0
botbuilder-core>=4.14
botbuilder-integration-aiohttp>=4.14
aiohttp>=3.9
pydantic>=2.5
python-dotenv>=1.0
jinja2>=3.1
apscheduler>=3.10
